+++ NEWS.md
@@ -1,5 +1,10 @@
-# moodef 1.0.0
+# moodef 1.1.0
 
-# moodef 0.1.0
+* Generate questions from an Excel file. 
+* Create an Excel file.
+* Read an Excel file.
+* Add name to all columns.
+
+# moodef 1.0.0
 
 * Initial CRAN submission.